refactor(core): type file for keriaNotificationService (#1166)
* refactor(core): type file for keriaNotificationService

* build: audit fix

* ci(gha): audit before tests and audit on main branch too

* ci(gha): npm ci before audit
