+++ src/core/agent/agent.types.ts
@@ -1,6 +1,5 @@
 import { SignifyClient } from "signify-ts";
 import { CoreEventEmitter } from "./event";
-import { OperationPendingRecordType } from "./records/operationPendingRecord.type";
 import { ConnectionHistoryType } from "./services/connectionService.types";
 
 enum ConnectionStatus {
@@ -124,17 +123,6 @@ interface AuthorizationRequestExn {
   e: { rpy: NotificationRpy; d: string };
 }
 
-interface KeriaNotification {
-  id: string;
-  createdAt: string;
-  a: Record<string, unknown>;
-  connectionId: string;
-  read: boolean;
-  groupReplied: boolean;
-  groupInitiatorPre?: string;
-  groupInitiator?: boolean;
-}
-
 enum OobiType {
   NORMAL = "NORMAL",
   MULTI_SIG_INITIATOR = "MULTI_SIG_INITIATOR",
@@ -147,11 +135,6 @@ type OobiScan =
       groupId: string;
       connection: ConnectionShortDetails;
     };
-
-interface KeriaNotificationMarker {
-  nextIndex: number;
-  lastNotificationId: string;
-}
 
 interface AgentServicesProps {
   signifyClient: SignifyClient;
@@ -169,43 +152,10 @@ interface AgentUrls {
   bootUrl: string;
 }
 
-enum NotificationRoute {
-  // "Real" notifications from KERIA
-  MultiSigIcp = "/multisig/icp",
-  MultiSigExn = "/multisig/exn",
-  MultiSigRpy = "/multisig/rpy",
-  ExnIpexApply = "/exn/ipex/apply",
-  ExnIpexOffer = "/exn/ipex/offer",
-  ExnIpexAgree = "/exn/ipex/agree",
-  ExnIpexGrant = "/exn/ipex/grant",
-  // Notifications from our wallet to give further feedback to the user
-  LocalAcdcRevoked = "/local/acdc/revoked",
-  LocalSign = "/local/signrequest",
-  LocalConfirmation = "/local/signconfirmation",
-  LocalInformation = "/local/signinformation",
-  LocalConnectInstructions = "/local/connectinstructions",
-}
-
-enum ExchangeRoute {
-  IpexAdmit = "/ipex/admit",
-  IpexGrant = "/ipex/grant",
-  IpexApply = "/ipex/apply",
-  IpexAgree = "/ipex/agree",
-  IpexOffer = "/ipex/offer",
-}
-
 interface BranAndMnemonic {
   bran: string;
   mnemonic: string;
 }
-
-type OperationCallback = ({
-  oid,
-  opType,
-}: {
-  oid: string;
-  opType: OperationPendingRecordType;
-}) => void;
 
 enum CreationStatus {
   PENDING = "PENDING",
@@ -220,14 +170,7 @@ export const OOBI_AGENT_ONLY_RE =
 export const DOOBI_RE = /^\/oobi\/(?<said>[^/]+)$/i;
 export const WOOBI_RE = /^\/\.well-known\/keri\/oobi\/(?<cid>[^/]+)$/;
 
-export {
-  ConnectionStatus,
-  MiscRecordId,
-  NotificationRoute,
-  ExchangeRoute,
-  OobiType,
-  CreationStatus,
-};
+export { ConnectionStatus, MiscRecordId, OobiType, CreationStatus };
 
 export type {
   ConnectionShortDetails,
@@ -235,9 +178,7 @@ export type {
   ConnectionNoteDetails,
   ConnectionNoteProps,
   ConnectionHistoryItem,
-  KeriaNotification,
   OobiScan,
-  KeriaNotificationMarker,
   AgentServicesProps,
   IdentifierResult,
   AgentUrls,
@@ -245,7 +186,6 @@ export type {
   ExnMessage,
   NotificationRpy,
   AuthorizationRequestExn,
-  OperationCallback,
   JSONValue,
   JSONObject,
 };
