+++ src/core/agent/records/notificationRecord.ts
@@ -1,5 +1,5 @@
 import { BaseRecord, Tags } from "../../storage/storage.types";
-import { NotificationRoute } from "../agent.types";
+import { NotificationRoute } from "../services/keriaNotificationService.types";
 import { LinkedRequest } from "./notificationRecord.types";
 
 interface NotificationRecordStorageProps {
