+++ src/core/agent/services/identifierService.ts
@@ -9,8 +9,8 @@ import {
   AgentServicesProps,
   IdentifierResult,
   MiscRecordId,
-  NotificationRoute,
 } from "../agent.types";
+import { NotificationRoute } from "./keriaNotificationService.types";
 import {
   IdentifierMetadataRecord,
   IdentifierMetadataRecordProps,
