+++ src/ui/__fixtures__/notificationsFix.ts
@@ -1,4 +1,4 @@
-import { KeriaNotification } from "../../core/agent/agent.types";
+import { KeriaNotification } from "../../core/agent/services/keriaNotificationService.types";
 
 const now = new Date(Date.now());
 const notificationsFix: KeriaNotification[] = [
