+++ src/ui/pages/Notifications/Notifications.test.tsx
@@ -13,7 +13,7 @@ import { credsFixAcdc } from "../../__fixtures__/credsFix";
 import { notificationsFix } from "../../__fixtures__/notificationsFix";
 import { NotificationFilters } from "./Notification.types";
 import { Notifications } from "./Notifications";
-import { NotificationRoute } from "../../../core/agent/agent.types";
+import { NotificationRoute } from "../../../core/agent/services/keriaNotificationService.types";
 import { NotificationItem } from "./NotificationItem";
 
 mockIonicReact();
