+++ src/ui/pages/Notifications/Notifications.tsx
@@ -2,10 +2,7 @@ import { IonList, useIonViewWillEnter } from "@ionic/react";
 import { useEffect, useMemo, useRef, useState } from "react";
 import { useHistory } from "react-router-dom";
 import { Agent } from "../../../core/agent/agent";
-import {
-  KeriaNotification,
-  NotificationRoute,
-} from "../../../core/agent/agent.types";
+import { KeriaNotification , NotificationRoute } from "../../../core/agent/services/keriaNotificationService.types";
 import { i18n } from "../../../i18n";
 import { TabsRoutePath } from "../../../routes/paths";
 import { useAppDispatch, useAppSelector } from "../../../store/hooks";
