+++ src/ui/pages/Notifications/components/EarlierNotification.types.ts
@@ -1,4 +1,4 @@
-import { KeriaNotification } from "../../../../core/agent/agent.types";
+import { KeriaNotification } from "../../../../core/agent/services/keriaNotificationService.types";
 
 interface EarlierNotificationProps {
   data: KeriaNotification[];
