+++ src/ui/pages/Notifications/components/NotificationOptionsModal.test.tsx
@@ -3,7 +3,7 @@ import { fireEvent, render, waitFor } from "@testing-library/react";
 import { ReactNode, act } from "react";
 import { Provider } from "react-redux";
 import configureStore from "redux-mock-store";
-import { KeriaNotification } from "../../../../core/agent/agent.types";
+import { KeriaNotification } from "../../../../core/agent/services/keriaNotificationService.types";
 import EN_TRANSLATIONS from "../../../../locales/en/en.json";
 import { TabsRoutePath } from "../../../../routes/paths";
 import {
