+++ src/ui/pages/Notifications/components/NotificationOptionsModal.types.ts
@@ -1,4 +1,4 @@
-import { KeriaNotification } from "../../../../core/agent/agent.types";
+import { KeriaNotification } from "../../../../core/agent/services/keriaNotificationService.types";
 
 interface NotificationOptionModalProps {
   optionsIsOpen: boolean;
